Support running just one packager for faster experiments

+++ bin/get-results.sh
@@ -10,32 +10,62 @@ resultsDir="${rootDir}/results"
 metroPort=9091
 haulPort=9092
 
+if [ $# -gt 0 ]
+then
+    do_metro=0
+    do_haul=0
+    for var in "$@"
+    do
+        if [ $var == "metro" ]
+        then
+            do_metro=1
+        elif [ $var == "haul" ]
+        then
+            do_haul=1
+        else
+            echo "Unknownt argument $var"
+            exit 1
+        fi
+    done
+else
+    do_metro=1
+    do_haul=1
+fi
+
 cd "${rootDir}"
 
-# Spawn metro server
-yarn start --port $metroPort &
-metroPid=$!
+if [ $do_metro -gt 0 ]
+then
+    # Spawn metro server
+    yarn start --port $metroPort &
+    metroPid=$!
 
-sleep 1
+    sleep 5
 
-# Spawn haul server
-yarn haul start --port $haulPort &
-haulPid=$!
+    # Get source from Metro
+    curl -so "${resultsDir}/metro-index.js" "http://localhost:${metroPort}/index.bundle?platform=ios&dev=true&minify=false"
 
-sleep 1
+    # Get source map from Metro
+    curl -s "http://localhost:${metroPort}/index.map?platform=ios&dev=true&minify=false" | jq . > "${resultsDir}/metro-map.json"
 
-# Get source from Metro
-curl -so "${resultsDir}/metro-index.js" "http://localhost:${metroPort}/index.bundle?platform=ios&dev=true&minify=false"
+    # Kill metro before we go
+    kill $metroPid
+fi
 
-# Get source map from Metro
-curl -s "http://localhost:${metroPort}/index.map?platform=ios&dev=true&minify=false" | jq . > "${resultsDir}/metro-map.json"
+if [ $do_haul -gt 0 ]
+then
+    # Spawn haul server
+    yarn haul start --port $haulPort &
+    haulPid=$!
 
-# Get source from Haul
-curl -so "${resultsDir}/haul-index.js" "http://localhost:${haulPort}/index.bundle?platform=ios&dev=true&minify=false"
+    sleep 5
 
-# Get source map from Haul
-curl -s "http://localhost:${haulPort}/index.ios.bundle.map" | jq . > "${resultsDir}/haul-map.json"
+    # Get source from Haul
+    curl -so "${resultsDir}/haul-index.js" "http://localhost:${haulPort}/index.bundle?platform=ios&dev=true&minify=false"
 
-# Stop metro and haul before we go
-kill $metroPid
-kill $haulPid
+    # Get source map from Haul
+    curl -s "http://localhost:${haulPort}/index.ios.bundle.map" | jq . > "${resultsDir}/haul-map.json"
+
+    # Stop haul before we go
+    kill $haulPid
+fi
